requirements.txt
Pillow>=10.0.0
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
